requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
asyncpg==0.29.0
attrs==25.4.0
bcrypt==4.0.1
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
distro==1.9.0
ecdsa==0.19.1
fastapi==0.128.0
filelock==3.20.2
frozenlist==1.8.0
fsspec==2025.12.0
greenlet==3.0.3
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface-hub==0.36.0
idna==3.11
importlib-metadata==8.7.1
jinja2==3.1.6
jiter==0.12.0
joblib==1.5.3
litellm==1.22.3
markupsafe==3.0.3
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numpy==2.4.1
openai==2.14.0
packaging==25.0
passlib==1.7.4
propcache==0.4.1
psycopg2-binary==2.9.11
pyasn1==0.6.1
pydantic==2.12.5
pydantic-core==2.41.5
pydantic-settings==2.12.0
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.21
pyyaml==6.0.3
redis==5.0.1
regex==2025.11.3
requests==2.31.0
rsa==4.9.1
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.0
sentence-transformers==5.2.0
setuptools==69.0.3
shellingham==1.5.4
six==1.17.0
sniffio==1.3.1
sqlalchemy==2.0.23
starlette==0.50.0
sympy==1.14.0
threadpoolctl==3.6.0
tiktoken==0.12.0
tokenizers==0.22.2
torch==2.10.0
tqdm==4.67.1
transformers==4.57.6
typer-slim==0.21.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uvicorn==0.40.0
yarl==1.22.0
zipp==3.23.0
